requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
crispy-bootstrap5==2026.3
dj-database-url==3.1.2
Django==6.0.3
django-cloudinary-storage==0.3.0
django-crispy-forms==2.6
django-unfold==0.83.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
